stp: rename iframe preview file
