+++ apps/full-site-editing/full-site-editing-plugin/starter-page-templates/page-template-modal/components/template-selector-preview.js
@@ -22,7 +22,7 @@ import { createBlock } from '@wordpress/blocks';
 /**
  * Internal dependencies
  */
-import BlockFramePreview from './template-preview-frame';
+import BlockFramePreview from './block-iframe-preview';
 
 const THRESHOLD_RESIZE = 300;
